+++ chatLayerSpecs/logic-events-conversation-updated.spec.ts
@@ -1,6 +1,6 @@
 import {
-    ComapiChatLogic
-} from "../chatLayer/src/chatLogic";
+    ComapiChatClient
+} from "../chatLayer/src/comapiChatClient";
 
 
 import {
@@ -155,12 +155,12 @@ describe("Chat Logic tests", () => {
             .withMessagePageSize(10)
             .withLazyLoadThreshold(10)
 
-        let chatLogic = new ComapiChatLogic(foundation, chatConfig);
+        let chatClient = new ComapiChatClient();
 
-        return chatLogic.initialise(chatConfig)
+        return chatClient._initialise(foundation, chatConfig)
             .then(result => {
                 expect(result).toBeDefined();
-                return chatLogic.getConversations();
+                return chatClient.messaging.getConversations();
             })
             .then(result => {
                 expect(result).toBeDefined();
