+++ chatLayerSpecs/logic-synchronise-events.spec.ts
@@ -1,6 +1,6 @@
 import {
-    ComapiChatLogic
-} from "../chatLayer/src/chatLogic";
+    ComapiChatClient
+} from "../chatLayer/src/comapiChatClient";
 
 
 import {
@@ -160,7 +160,7 @@ describe("Chat Logic tests", () => {
             .withMessagePageSize(10)
             .withLazyLoadThreshold(10)
 
-        let chatLogic = new ComapiChatLogic(foundation, chatConfig);
+        let chatClient = new ComapiChatClient();
 
         store.createConversation({
             id: "D35A13DF-6876-4CC8-BA70-841B45A0003C",
@@ -183,11 +183,11 @@ describe("Chat Logic tests", () => {
                 });
             })
             .then(result => {
-                return chatLogic.initialise(chatConfig)
+                return chatClient._initialise(foundation, chatConfig)
             })
             .then(result => {
                 expect(result).toBeDefined();
-                return chatLogic.getConversations();
+                return chatClient.messaging.getConversations();
             })
             .then(result => {
                 expect(result).toBeDefined();
