+++ package.json
@@ -47,6 +47,6 @@
     "webpack": "^1.13.3",
     "ws": "^1.1.1"
   },
-  "build": "14",
+  "build": "16",
   "license": "MIT"
 }
